requirements.txt
environs==9.5.0
requests==2.28.2
python-telegram-bot==13.0
django==4.2
APScheduler==3.6.3